Add a driver for the Cirrus Logic wm9884 codec
This code has been uncereminiously ganked from U-Boot. All it really
does is configure the codec for I2S mode and enables the headphone
output. This requires some twiddling via I2C. The ST Micro STM32F746
Discovery board has one of these chips on it. It's likely we'll use
a different chip for the projection board, but we need this for testing
and development in the meantime.