Update the video player and switch to libjpeg-turbo
The TJPGD decoder is nice, but libjpeg-turbo is a bit faster,
particularly with some complex and higher quality frames. It's
not a huge difference (there's no hardware decode assist or
anything), but it's noticeable. (In one case playback got a
little stuttery due to the amount of time it took to decode
some frames. That doesn't happen with libjpeg-turbo.) There is
also a significant performance gain when libjpeg-turbo is compiled
with -O3.

The trade off is that libjpeg-turbo is also noticeably larger in
terms of compiled code size. However we still have most of the
on-board flash space left, so I don't think this will be a big
issue.

I also added a temporary cmd-video.c utility to exercise the
video player. (Note that there isn't a separate app thread yet
so if you use this command it will block the shell until the video
finishes playing.)