Use less memory in the video player
Now that we're using so much RAM for double-buffered video, we
should avoid being a pig in the video player. We don't need to
buffer the entire frame in RAM just to copy it to a frame layer
buffer later. We can do it a couple scanlines at a time.