Make some more video player performance improvements
We always know exactly what buffers the jpeg library is going to
allocate to decode a frame, so we can rig the jmemchibios.c to
reserve exactly those buffers. This avoids some overhead in calling
malloc()/free() for every frame.

Also, using the internal SRAM for work buffers seems to yield
better performance (possibly due to the external SDRAM not being
quite as fast).