Adjust the priorities of the video and music apps
Orchard apps usually run at a very low priority, lower that the GTimer
thread in particular. When we're using the GTimer thread to poll for
touch events, it can sometimes preempt the music player app and
introduce some gaps during playback. This was also an issue in the
DC27 design and I used the same solution there, which was to bump up
the priority of the music player thread.