Add intro app and preliminary radio driver code
The intro app allows us to customize the badge start-up animations/sounds/etc
without needing to modify main.c any further.

The SX1262 driver is designed to work with the mBED SX1262MB2CAS shield,
which fits on the back of the discovery board. Right now it just has
basic code to turn the radio on and send a few packets, with interrupt
handling. This will need more refinement, but I want to start tracking
the code in source management.