Fix a few more network issues
In order to make sure we correctly sync the cache when tranferring
data to the radio, we need to make sure the transfer buffer is aligned
on a cache line boundary. Otherwise we may send corrupted packet data.

We can now tell lwIP to use the SDRAM heap memory (via malloc()/free())
for its data buffers instead of relying on pools allocated from the
BSS. This really fixes the ability to send larger IP datagrams. (Before
we were limited to 16 RX buffers due to the PBUF pool only having 16
pbufs in it by default.)

Also, the ping command now supports a packet count argument.