Fix the SAI PLL initialization
We need to tell ChibiOS that the LCD is present before it will perform
the SAI PLL setup. Also, we need to adjust some of the timing values
for the screen on the discovery board. This may need some additional
fine tuning once we start using the production 320x240 screen.