Clean up display configuration a little
Fix it so that uGFX things the screen is actually 320x240
pixels in size, which matches the expected resolution of
the screen that will be on the production badges.

Note that part of the driver still uses the 480x272 resolution
but that's just because it affects the clock/pixel timings, and
those still have to match the display on the STM32F746 Discovery
board.

Also, try to use gdispGetWidth()/gdispGetHeight() instead of
hard-coded constants where applicable, and fix up the scrolling
code to get rid of all the the 480x272 vs 320x240 fixups.