Fix the ChibiOS DMA2D and LCD controller drivers
The LCD and DMA2D drivers don't enable their clocks for low power
mode. We want that since we use WFIs when idle to conserve power.

Also, the LCD controller has code in it to set the clock which is
a) specific to the STM32F4xx controller (the registers are different
for the STM32F7xx) and b) redundant since the STM32F7xx clock init
code is already set up to initialize the clock divisors correctly.