Put thread stacks in the heap instead of the BSS.
With the STM32F746, ChibiOS uses the DTCM RAM, which is the first 64KB
at address 0x20000000 as the BSS. The chip actually has 256KB more SRAM,
but the DTCM RAM is special in that it's not cached. ChibiOS actually
takes advantage of this. Many of the STM32 peripherals use DMA channels
for I/O, and the STM32F746 has a data cache, so you need to make sure
to keep the cache synrhonized correctly when doing DMA transfers.
But cache synchonization is complicated, so ChibiOS tries to dodge the
issue: it expects that people are going to use static buffers as DMA
targets, and static buffers are allocated from the BSS. If the BSS is
also uncached, then you you don't need to worry about cache synchronization.

Also, thread stacks are allocated from the BSS too, if you use
chThdCreateStatic().

But there is only 64KB of DTCM RAM, and with a complicated enough
application (like ours) you need to be frugal with its use.

So to avoid using up too much of the BSS, we now use chThdCreateFromHeap()
for all our own threads instead. Also, some larger buffers (like the audio
playback buffer, which is several KB in size) are allocated from the heap
instead of from the BSS too.

For one thing, this allows us to switch the JPEG library back to the
previous strategy of using pre-allocated static buffers for its internal
memory allocations. While the DTCM RAM is uncached, using it for the
JPEG library's internal buffers seems to provide noticeably better
performance when playing back video. I'd like to keep the performance
as good as possible, but I also want to conserve BSS space, so in
exchange for putting these buffers back in the BSS, we move all the
thread stacks and other large buffers to the heap.

One side effect from this was that the voltage and temperature checking
routines which use the ADC started displaying bizarre results. This was
due to the sample buffers now being allocated from the cached SRAM instead
of the uncached DTCM RAM. A fix was made to the ChibiOS STM32 ADCv2 driver
to address this in a previous commit.

A change has also been made to the radio driver to reduce how often
it actually generates interrupts to wake up the CPU. We don't necessarily
need to be notified if the receiver detects a CRC or header error (it
will keep listening for packets without us having to prod it), so we
don't need those in the dio1 mask.

Lastly, With GCC 10.1, it seems the best performance can be achieved
by going back to -Os again, so the Makefile has been updated accordingly.