Fix 64-bit cleanliness bug
Casting a native pointer to (uint32) is technically the wrong thing
to do when calculating alignment fixups. You should really use
uintptr_t, otherwise the code is not 64-bit clean. Technically
this doesn't here because we intend this code to run on a Cortex-M7
core which is only 32-bit, but I want this bug documented somewhere
in case anybody else ever wonders why Nofrendo segfaults when you
try to compile and run it on their 64-bit machine (like it did for
me with FreeBSD/amd64).