Add the NES emulator to the build
A small NES emulator app has been added that launches the emulator.
Note: right now the emulator runs and stays in attract mode, but
there's no way to exit due to lack of inputs. I'm still not
sure what to do about this since the Discovery board has just
one input button.

Also added three simple games and an icon for the emulator app.

Testing showed that the Orchard app stack was too small for the
Pac Man game, so I increased it to 4KB. We can get away with this
since we have plenty of available RAM.