Fix static variable bug in nofrendo
The nesinput.c module has a static variable called active_entries which is
assumed to be initialized to 0 each time the program is run. But here
in RTOS land, we need to do that manually. The static initialization
only counts when the system is booted.

Without this fix, you can only run the NES emulator about 16 times
before it crashes.