Make a first attempt at implementing sound for the NES emulator.
This works, but the output sounds scratchy and I'm not sure why.
The same code compiled on my desktop sounds fine (with the only
major difference being that my desktop is obviously much faster
than the STM32F746).

Also, switch to using the high resolution hardware timer for driving
the emulator instead of the virtual timer.

The NES emulator generates mono sound. For the time being I
added a function to switch from stereo to mono, however this may
be removed later as I doubt we'll have more than one speaker on
the badge anyway, so the default will have to be changed to mono
once the actual boards are built.