Fix a subtle bug
The sprite initialization code assumes that there is a NULL entry at
the end of the sprnames[] array. Well, there actually isn't. Sometimes
you get lucky if it happens that some other data happens to end up
positioned right after the sprnames[] array and that data is zeroed.
That seems to be the case most of the time, unless you build with LTO.

If you don't have a NULL entry right at the end of the sprnames[] array,
the calculation in R_InitSpriteDefs() that figures out the size of the
sprite namelist gets confused and can run off the end of the list.
We correct this by forcing a dummy entry into the end of the array which
is initialized to NULL.