Fix another obscure bug that breaks Doom demo playback
Pop quiz hot shot: is char signed or unsigned?

The answer is: it depends on the platform.

On x86 it usually defaults to signed, and it turns out that there
are places in the Doom code which depend on this assumption, notably
the command processing code. The ticcmd_t structure defines the
forwardmove and sidemove members to be char, and the data read
back from the demo stream uses the convention that the values
are signed. However on ARM, the default for char is usually unsigned.
This means that a value of "-4" for forwardmove in the demo stream
may be misinterpreted as "252" on ARM systems (both are 0xFC).

So remember kids: if you want chars to be treated as signed, you
_must_ declare them as "signed char" otherwise your code isn't
truly portable.

We could just change forwardmove and sidemove to signed char to
fix this, but there may be other abuses like this lurking in the
Doom code, so to save some aggravation we force the compiler to
assume signed chars instead of unsigned for consistency with x86.