Being over a couple of uGFX fixes from previous designs
[...]
The GetMouseReading() function has a local GMouseReading structure
which is stored on the stack, however it does not initialize this
structure before attempting to use it. This results in the r.buttons
sometimes containing bogus values which causes some strange behavior
of the mouse button event handling logic.

The structure is now zeroed before being used.
[...]

[...]
Add support for both BUTTON_UP and BUTTON_DOWN events
[...]