Adjust the dialer tone loop a little
We need to yield the CPU for a tick after starting samples playing
otherwise we might not give the gTimer thread a chance to run. Currently
we rely on it to signal when touch panel events occur, and if we
starve it out we might get stuck playing a tone forever.