Fix the dialer app for real this time
The stuck key problem occurs because the orchard uGFX callback
only allows one event at a time to be posted. The previous event
must be cleared before the next can be accepted. The GTimer thread
will block waiting for an existing event to be completed before
submitting the next. This is fine normally, but the dialer app
calls the tonePlay() routine in its event handler, which keeps
running until the user releases the dial button. This means no
events can be processed until the tonePlay() function returns.

To get around this, we disconnect the uGFX callback from the
uGFX listener queue before calling tonePlay() and re-attach it afterwards.
That way we never get stuck because the GTimer thread never calls
the function that blocks.