Adjust the SAI clock configuration
I noticed that when running the dialer app, there would sometimes
be some graphics glitches when pressing some of the buttons. I
surmised this was due to glitching in the LCD controller clock.
Previously I had the SAI2 module connected to the I2S PLL. I
switched it to using the SAI PLL, and now the glitching seems
to have gone away.

The SAI PLL has two output divisors: one feeds the LCD controller
and the other feeds the SAI blocks. You would think that using
separate PLLs for the LCD and SAI blocks would be more stable
rather than less, but apparently that's not the case. I don't really
understand why the glitching occurs, but it doesn't happen anymore,
so I'm going to call it good.

Fortunately it was still possible to provide a 16MHz input clock
to the SAI2 block by adjusting the PLLSAIDIVQ divisor while still
retaining the 192MHz SAI PLL frequency (and without changing the
LCD clock frequency).