Finally fix the glitchy video problem once and for all.
I still don't fully understand this problem, but fixing it required
creating a mapping ifor Bank 1 of the FSMC controller with the MPU.
The mapping actually makes the region inaccessible by the CPU and
sets its attributes to uncached. I don't know why this is necessary,
but it does seem to fully fix the problem.