Add the first step towards power management
We can now change the CPU speed on the fly. Default is 216MHz, but
we can set it to 108MHz (medium speed) or 54MHz (slow speed).

The app-badge.c app will put the CPU in slow speed when it starts,
since it does next to nothing, and will restore the speed to normal
when it exits. Hopefully this will reduce power consumption when
the badge is idle. According to the datasheet, at 216MHz the CPU
can consume up to 230mA, but at 54MHz consumption may be around
65mA.

Note: the timer frequencies are affected when the CPU clock is
slowed down. I may need to address this later.