Add clock compensation for CPU speed changes
We delibrately choose slower APB1 and APB2 clock settings so that
we can adjust them when changing the CPU speed. This preserves the
speed for the system clock timer, so that time doesn't run slower
when we reduce the CPU speed. We also make some adjustments to the
SPI and I2C clocks (fortunately we can use their prescalers to
keep the fixups small).

System behavior at full CPU speed is unchanged.