Add support for enabling SDRAM power down during sleep
The badge_lpidle_enable() function will make it so that the SDRAM
controller clock will be disabled when entering sleep mode. The
plan is for the home screen app to do this to save power when the
badge is idle. If badge_sleep_enable() is called, the CPU will
do a WFI in the idle loop, which will put the core to sleep. If
badge_lpidle_enable() is also called, then the SDRAM controller
and LTCD clocks will be stopped as well (and a self-refresh
command will be sent to the SDRAM).