Fix up the ILI9341 driver a little
The existing ILI9341 driver uses some slightly dodgy logic. You really
need to set/clear the chip select pin in between commands, especially
if you intend to switch between write and read commands. We don't use
this driver yet, but we will eventually in the production design, so
I'd like it to work right. This was tested with the STM32F429 Discovery
board.