Add the MAC component to save some power
The STM32F746 discover board has an ethernet port on it. We don't
actually want to use it, but the 10/100 PHY that's attached to it
defaults to being enabled at power up. If we include the STM32
ethernet driver component, even if we don't use it, the macInit()
function will be called via halInit(), and that will cause the driver
to set the "power down" bit in the PHY control register, which
saves us about 40mA of current consumption.