Create a simple double buffer API
Since Doom and the NES emulator use basically the same logic for doing
double-buffered graphics, and since we might want to use the same
logic later in other apps, we may as well create some common routines
for handling the setup, teardown and blitting.

Note that the video player uses some special tricks for drawing,
so it can't use the common blit routine. But it can use the common
setup/teardown functions.