Fix a 23-year old Doom bug.
When you exit Doom, it taunts you with a message that's randomly
selected from the endmsg[] array in dstrings.c. This message is
rendered on screen using the bloody Doom font along with a "press
y to quit" prompt. The messages are broken up with linefeed characters
so that they fit neatly in the center of the display.

But there's a problem: two of the strings in the array are missing
commas after them ("go ahead and leave, see if i care" and "you're
lucky I don't smack you for thinking about leaving."). This means that
instead of being treated as separate array entries, the C compiler
concatenates them together with the string that follows them. This causes
Doom to print both of them together (and makes the list appear shorter
than it really is).

But even worse, it creates multi-line messages where one of the lines
is too long to fit entirely on the screen; when Doom renders, the
message, this causes the range check in V_DrawPatch() to complain that
it's drawing outside the screen buffer bounds. Sometimes this will also
lead to display corruption.

For a minute I thought maybe I butchered this file myself on accident,
but I checked the original zip file