Disable CPU speed changes in launcher
We bought some new STM32F746 Discovery boards, and the screens on them
behave a little differently than the first batch: the display goes white
immediately as soon as the LCD clock is disabled, which makes the screen
flash whenever the CPU speed is changed.

This may not be the behavior of the production hardware, but it's very
distracting so for now I've turned it off. (The intent was to save power
when running in the launcher, but we don't end up sitting in the launcher
very long anyway, so maybe this was overly aggressive.)