Add first cut of firmware update utility
This uses basically the same scheme as earlier badges except now
we're using the ChibiOS "os-less" configuration to take advantage of
using the ChibiOS drivers for the STM32F746 chip, notably the SD/MMC
driver.

To simplify things, we do a mass erase of the internal flash and then
reprogram it. If flashing fails, the board will be bricked, but you can
unbrick it either with JTAG/SWD or by using the internal DFU ROM and
dfu-utils.