Make I_AllocLow() return an aligned block of memory.
The save game feature uses an ugly cheat where it (ab)uses the
memory for screen[1] as a save game buffer. This buffer is then
written to disk. It helps if this buffer is aligned on a cache
line boundary (at the very least) otherwise sometimes the game
saves can be corrupted.