Add virtual keyboard support
This commit adds support for using a special program on a host
system to relay keyboard events to the badge using the serial
console. This works over either the actual UART or the virtual
USB com port. Once the program is running and the user hovers
their mouse over the window, any key up/down events will be
sent to the badge, captured, and transformed into joypad
events. The Doom and NES emulator apps can also acquire events
direct from the capture queue to simulated keyboard input.

This allows the user to play Doom on the badge, and allows the NES
emulator to work on the prototype hardware which doesn't have
physical inputs yet.