Move ethernet buffers out of the .bss section
The STM32 ethernet driver has a couple of big buffers used for packets
and descriptors. We're not using the ethernet port, but we include the
driver just so it will shut off the power to the PHY chip on the
Discovery board. Unfortunately this means including the driver bloats
the .bss section a bit due to the ethernet buffers. To mitigate this,
the .eth section is now part of .ram7 (the Doom BSS) instead of .ram3.