Adjust ULPI PHY control code a little
When turning off the VBUS supply, clear both the DrvVbus and
DrvVbusExternal bits in the OTG control register. These two bits
are ORed together to control the external 5V supply. (The new USB
host driver code that arrived with ChibiOS 20.3.3 sets the
USB_OTG_GUSBCFG_ULPIEVBUSD bit in the GUSBCFG register in the
host controller which forces DrvVbus to be 1 along with
DrvVbusExternal.)