Fix another Doom bug
The WAD code does a malloc() of 1 byte for the lumpinfo buffer which is
then used with realloc() latter. This confuses the heap allocatoe code
in newlib though: the single byte is allocated within the heap at a
point that is not far from the bottom of the Doom thread's stack.
When the realloc() occurs, realloc() wants to copy the data from
the original buffer (which is really just 1 byte) to the new buffer.
It does this using the new buffer's size. However the distance between
the old buffer of 1 byte and the start of the Doom stack is less than
that size. Since the bottom of the Doom stack has a guard page there,
this copy triggers a page fault.

To avoid this, we just initialize lumpinfo to NULL. This way realloc()
will behave like malloc() and just allocate a new block without trying
to copy data from an old block.