Clean up and add support for GFSK mode
The SX1262 now supports GFSK mode as well as LoRa mode. With GFSK
mode, we can get speeds up to 300Kbps (which is better than the
max 21Kbps you can get with LoRa).

Also, the driver is now no longer of the ethernet class. There's
really not much point to doing it that way: it just means we
chew up 14 extra bytes in the packet for the frame header and
use another timer in the LWIP stack.