Implement dynamic timers in LWIP
The LWIP stack uses a cyclic timer scheme which polls for timer
events periodically. The problem with this is that it keeps triggering
even if there aren't any timers actually pending, which causes the CPU
to wake up a lot for no good reason.

Now the timers are dynamic: once they expire, they're removed from
the timer list and terminated.