Make radio mode and TX power config options
The user can now save the default radio mode and TX power in the
configuration file on the SD card. The factory default is still
LoRa mode and 14dBm. But the user can select high power (22dBm) and
GFSK mode if they want now. The latter is really only useful for
testing.