Add early support for radio shouting
Since we can do broadcasts again, the shout app from the first
badge design has been resurrected. Also added is the basic framework
for notifications. Right now the shout notification is the only one
that works. Eventually there should be a game challenge notification.

The intent for some apps is that when a notification is sent, the
sender will specify a new radio channel to use. The two badges can
then switch to that channel to avoid talking over other badges that
may be nearby. This part still needs work, but should be possible
with the radio driver the way it's implemented now.