Work around a possible TX lockup
It seems like sometimes the radio can stop generating interrupts during
periods of heavy TX traffic (like when playing networked Doom). As a
workaround (until I can think of something better), we set a timer
on transmit for a few milliseconds and reset the chip if we don't
get a TX completion interrupt before the timeout expires.

There may be a less heavy-handed way to recover, but I'm not sure
what that might be yet.

Also, the GFSK bitrate has been increased to 350Kbps.