Modify the virtual keyboard capture mechanism a little
We need to avoid a conflict with the badge_console.c module here.
Uptating the console handles all the time doesn't really work:
it can lead to confusion of the method tables and potentially
a crash. (This is more evident now that I've been playing Doom
more often to test the network support.)

I may need to apply the same logic to the badge_console.c code
in a similar way later. For now this fixes the crashes.