Silence the last Doom warning
When using LTO, the compiler complains about a type mismatch due to
some of the action function pointers taking two pointer arguments instead
of just one. There's no easy way to fix this without significantly
changing a fair bit of code, and the warning is harmless (things still
seem to work ok anyway) so to get around this the build.mk file
for Doom will now add the -Wno-lto-type-mismatch option to the build
flags to silence the warning.