Check in initial support for networked Doom
- Badge finder module supports accept, decline and Doom request events

- Badge finder module now supports routines for changing radio channel
  and mode

- A table of 52 channels has been added which covers the 902MHz to
  928MHz ISM band

- A Net Doom app has been added

When running the Net Doom app, you'll be given a list of peers and
you can select one to challenge. They will have 10 seconds to accept
or decline the challenge. If the peer accepts, both badges will switch
to a randomly selected channel and change to GFSK mode, then launch
a network Doom session.

Currently Doom always starts with medium difficulty and at episode 1,
map 1. The Net Doom app will have a UI added to it to let the requesting
user change this, and optionally select deathmatch mode. This will come
later.