Add music support for Doom!
The WildMIDI library is now tied into Doom and can play the music
for each map, and the intro and intermission scenes.

Also, there was a bug in the i_sound.c module which was carried
over into i_sound_badge.c: the mixbuffer buffer was twice as large
as it needed to be. Previously said this:

// It is 2 for 16bit, and 2 for two channels.

But mixbuffer is declared as:

signed short    mixbuffer[MIXBUFFERSIZE];

Since this is an array of shorts, the "2 for 16bit" part is already
implied. That means BUFMUL only needs to be 2.

Also, in i_system.c, the amount of RAM that Doom allocates for its
internal allocator has been reduced from 6MB to 5MB. This is necessary
as WildMIDI chews up a lot of RAM. Fortunately this doesn't appear to
affect the game: I tested both the shareware and commercial Doom WAD
files, and they seem to run fine.