Change malloc() override strategies
There was still a problem with the previous approach. Rather than
relying on the linker using the right object code, we can just
always point the _foo_r() functions at our overide versions. We
now build the Doug Lea malloc with the "dl" prefix on all the
routines to keep them from clashing with the symbols names in
newlib.