Fix race condition in lpidle code
I noticed that sometimes when sitting overnight in lpidle mode,
one of the boards would get stuck with interrupts disabled, i.e. PRIMASK
register == 1. Forcing PRIMASK to 0 would unjam it. I think the problem
is a race condition in the badge_idle() routine. We need to guard
all of the code with __disable_irq()/__enable_irq(). I'll need to
keep an eye on this to make sure the problem doesn't continue to
occur, as I'm still not 100% sure I understand the failure condition.
But with this change, the board ran all night without failing.