Improve performance of Doom FixedDiv2() function
Originally Doom used assembly language routines for FixedMul() and
FixedDiv2(). The linuxdoom code uses C equivalents. It turns out
the FixedMul() function compiles to only 3 ARM instructions and is
pretty fast, but FixedDiv2() ends up using floating point intrinsics
and is fairly slow. This is mostly because it uses doubles instead
of floats, and ST Micro configured the STM32F746's FPU to only
support single precision instead of double precision.

Changing the doubles to floats in the C code improves things considerably,
but we can do better. There is now a hand-written assembly version of
FixedDiv2() that uses the FPU and which takes only 6 instructions.

Also, badge_fault.S has been update to correctly specify the CPU
and FPU types. (It still had the old info for the Cortex-M4 that
was in the DC27 badge.)