Add support for running from RAM
If BOOT_FROM_RAM is set to yes in the Makefile, the badge.bin image
can be loaded and run in RAM using U-Boot (or some other loader).
We may have some trouble getting the version of the STM32F746 with
1MB of flash. If so, we may have to settle for the 512KB flash version.
Our image size is already over that limit, so the only alternative
is to run from RAM.

The current configuration runs from the 8MB flash on the Discovery
dev board, but with any luck we'll be able to get 16MB for the actual
production board and we can adjust things a little. One adjustment
is to reduce Doom's zone allocator region to 4MB. This is necessary
to run Doom when music support is enabled. The shareware Doom WAD
seems to still play just fine like this, but if we can get a 16MB
SDRAM part, we can put it back to the original 6MB size and have
room to spare.