Don't disable MPU in trap handler when running from RAM
By default, the address range containing the SDRAM window is maked by
the Cortex-M7 CPU as non-executable (NX). This is overridden by the MPU
so that we can run code out of RAM. Turning the MPU off un-does this,
which causes us to die screaming. So avoid doing that in the run-from-RAM
case.