Try to avoid race conditions in radio driver
The design of the SX1262 chip is such that it can be difficult to avoid
race conditions between the RX and TX paths. To try and mitigate this,
we now use a strategy similar to that of the radio driver in the DC25
badge (which as also made by SemTech) which has the transmit routine
check for completions synchronously and the RX handler triggered
asynchronously.

This seems to result in better behavior when playing networked Doom
games. It may be possible to remove the -dup 2 argument now, but I
will leave that for when I update the netdoom app to allow specifying
game parameters.