Fix a bug in the WildMidi MUS parser code
MUS songs have drums (percussion) on channel 9 while real MIDI
songs have drums on channel 15. (This is mainly because MUS files only
have 9 channels.) The WildMidi library deals with this by swapping
channels 9 and 15 in the MUS data while parsing it. Unfortunately it
doesn't bother to swap them back when it's done.

In Doom, if a song is played twice without quitting the game, this
bug causes the second playback to sound wrong. (WildMidi is actually
modifying the cached MUS lump data.)

To correct this, we now un-do the fixup after we finish parsing a
MUS event.