Try to improve performance of the Nintendo emulator
Nofrendo has a vid_blit() function which it uses to blit bitmaps into
the display buffer. We can speed this up (supposedly) by using the DMA2D
engine as a hardware blitter. I'm not certain just how much this
improves things. In theory it should save on CPU cycles, but there's
no noticeable change in gameplay in the emulator. But this is easy
enough to revert later if it turns out to be waste of time.