Fix double-buffered graphics handling
Okay, I'm an idiot. Who doesn't understand how to do XORing correctly.
All this time when I thought the idesDoubleBufferBlit() was swapping
between display layers on ever call, it actually wasn't. So the video
player was using double-buffered graphics correctly (since it handles
this internally), but Doom and the Nintendo emulator were not.

I'll be over there, feeling stupid.