Add PLL setup to allow SAI sample rate of 11025KHz (almost)
It turns out we can use either the I2S or SAI PLLs as clock sources
for the SAI2 block. With careful programming, we can get SAI2
sample rate to 11029.412KHz, which is much closer to the 11025KHz
that Doom uses for its sound effects than the 10416.666KHz we were
using before.