Improve graphics performance for Doom
It turns out that we can get a big boost in graphics performance if
we put the Doom foreground and background screen buffer in the CPU's
internal SRAM instead of keeping it in the SDRAM. Doom draws to these
buffers a lot, especially the foreground buffer, and the SRAM faster
access times than the external SDRAM.

With this change, we can reduce the number of samples we play during
each loop, and this in turn makes the graphics _much_ faster/smoother.
This is likely about as good as it would have looked on a PC (especially
now that I also fixed the double-buffered graphics).

In the process, this also fixes a wasted memory allocation. It turns
out the V_Init() function already allocates all 4 primary screens
including screen[0], but I was allocating screen[0] from the SDRAM
heap again in I_InitGraphics(). This has been corrected now, which
avoids wasting 64000 bytes. (The memory was not leaked, but Doom
was allocating it for no reason.)