Update memory usage some more
Avoid using the ChibiOS heap in the uGFX library and the ChibiOS
lwip shim routines. This can lead to fragmentation when apps use
the heap.

Also, in Doom, just put the foreground buffer in the on-board SRAM.
There doesn't seem to be much benefit to putting the background
buffer there too, and we can't do it for the case where we also
put the initialized data in SRAM (there's not enough space available).

Finally, when building for BOOT_FROM_RAM put the initialized data
segment into SDRAM.