Restore original Doom code where appropriate
I'd like to be able to build as much of this code as I can on UNIX
(since it has many bug fixes in it). Since the i_xxx.c files are
system-specific, we now have all our own i_xxx_badge.c files, and
the originals are restored.

Also, we now have a -DBADGEDOOM macro defined by build,mk in order
to special-case any changed in other files.