Fix another Doom bug
When Doom does a screen wipe, it doesn't call the I_UpdateSound()
or I_SubmitSound() functions, even though it probably should. This
means when, for example, you reach the end of a level and press the
exit button, the screen wipes first and then you hear the "thonk"
of the button being pushed.

Note that even with this fix in place, there is still a bit of a
delay at this point, because we also transition between songs --
the intermission screen has its own song. The delay is caused by
the WildMidi library doing I/O to the SD card to load the necessary
patch files to render the song. There's not much that can be done
about this.