Fix another Doom bug, this one of my own making
finalecount must be an int. Previously it was changed to size_t
to avoid a compiler warning, but that breaks the finale behavior.
When the finale text is displayed, the calculation that tries
to figure out how much text to draw in F_TextWrite() goes bananas
and draws the whole text in the first frame, which is visible right
after the screen wipe. It also breaks the bunny scroll at the end
of the last map in the registered version of the game.