Fix another Doom bug
The code EV_BuildStairs() does not fully initialize its floormove_t
structures after allocating them. This can later result in sanity
check failing somewhere in the code that draws the textures on the
steps and makes the game exit. I tripped this once while playing.

The fix I used incorporates some of the corrections to fix this
bug in the Boom code (which is a port of the LinuxDoom sources
back to MS-DOS). Rising staircases still seem to work correctly
and the game plays ok, so I think this is sufficient.