Fix a Doom sound bug (of my own making)
There is exactly one place in Doom that I can find where it plays a
song without looping and then lets it reach the end, and that's at
the finale of the game. If you have the commercial/registered WAD
file, Doom plays a song during the bunny scroll which then ends.
However the music playing code doesn't handle this right: if the
song ends, we need to stop doing any of the mixing steps in
I_UpdateSound(), otherwise we end up playing a batch of noisy
samples over and over again (which is very jarring).