Fix a Doom II bug (no, really)
The IDCLEVxx cheat code for warping to different maps/episodes
doesn't work for gamemode == commercial (Doom II). It forces epsd to 0,
but then later there's a sanity check for epsd < 1. It turns out that
Doom II was formatted such that it had 34 maps in 1 episode, so the
episode number really should always be forced to 1, not 0.