Properly initialize menu items
For Doom II (commercial) games, the menu code changes some of the
default menu structure values. We need to make sure this are
always initialized to defaults when the game starts. A side effect
of this if we don't is that each time to restart the program when
playing Doom II, the main menu is drawn lower and lower on the
screen.