Add improved support for PWADs
There exist certain Doom mods based on PWAD files which don't work
with vanilla doom for a number of reasons, notably the Ultimate
Simpsons PWAD. But I kind of want it to work, so I've made some
patches to allow it.

These patches allow a PWAD to only partially remap some of the
game sprites; with the original implementation, remapping sprites
was basically an all or nothing proposition. Now partial remappings
work.

Also, we now have the ability to nominally parse SNDINFO lumps
which fixes certain sound glitches in the Simpsons PWAD.