Fix building of this program on newer FreeBSD versions
We need to know both the object format and architecture type in
order to get objcopy to generate the right kind of image, so
we need to figure that out from an existing ELF object too.

Also, use the main.o we just compiled for sampling the host
object format and architecture type instead of /bin/ls, since
we always know it will be there.