Port the virtkeys utility to Windows.
The virtkeys program can now be compiled on Windows using the MSYS 2.0
environment and MINGW compiler. This means we now support all major
platforms (FreeBSD, Linux, MacOS, Windows).

The documentation will need to be updated to show how to compile on MacOS.
I did this once before on John's laptop, but I need to double check how I
did it. (I think it was done using the SDL2 library installed from brew.)
Also, we need to test on an M1 Mac to make sure we can build an ARM image.