Update the audio parameters a little
For best results, it's best if the time it takes to play a batch of
audio samples is just slightly longer than the Doom tick time. The
tick rate is 35 times a second, meaning we have 1/35th of a second
to handle each frame, including the time it takes to sample user
inputs, update game state, generate and render the next video frame,
and generate the next round of audio samples. We want to do all of
that while the current batch of audio samples is playing. Then, once
we finish all that work, we pause until the current batch of samples
finishes and immediately queue up the next batch.

1/35th of a second is .02857 seconds.
This is also equal to the time it takes to play 315 samples at
11025Hz.

However we don't want the playback time to be exactly equal to the
frame rate: it should be slightly longer. Also, our actual playback
rate is 11029Hz, not 11025Hz.

316 samples played at 11029Hz is .02865, which is just slightly
longer. So we use these values for now.