Fix support for using a .doomrc file.
Previously I had disabled support for this, along with some other code
in d_main.c, because it required using malloc() and Doom is very bad
about releasing malloc()ed memory on exit. In other words, it
introduces memory leaks. However using a .doomrc file is useful in
that it lets the user save their settings between games, including
things like screen size selection and gamma correction.

So I fixed the memory leak problems and .doomrc support works now.
Also, Doom now actually uses the DOOMWADDIR and HOME environment
variables for paths.

We can't actually use a file called .doomrc though because with a FAT
filesystem, this name is not valid. Instead in app-doom.c, we override
the default file name and call it just doomrc instead. (This avoids
the need to change the original code.)

Also, some memory leaks needed to be fixed:

- Doom malloc()s strings in M_LoadDefaults(), but never free()s them.
This is now done in M_SaveDefaults() on exit.
- Doom malloc()s strings for WAD pathames in IdentifyVersion(), but
doesn't release them when exiting the function. This is also fixed.
- D_AddFile() malloc()s strings for WAD pathnames *again* and saves
them in the wadfiles[] array, but these are also never free()d. This
is fixed in i_system_badge.c.